Create stuck_in_the_mud_game.py
The game begins with the player 'rolling 5 dice.' 
Then the player gets a score of the combined numbers of the dice excluding 2's and 5's. 
Any 2's or 5's rolled results in dice "stuck in the mud." 
So say you roll a [1, 3, 4, 5, 2] the score should be: 8. 
Then the player rolls the remaining 3 dice, again you rolled [6,3,2] the score should then be 17. 
So on and so forth until the player runs out of dice and the code returns a final score.