Create fizz_buzz_three_ways.py
The FizzBuzz question is a simple programming challenge that is often used as a screening question in technical interviews. The challenge is to write a program that prints the numbers from 1 to 100, but with the following modifications:

For multiples of 3, print "Fizz" instead of the number.
For multiples of 5, print "Buzz" instead of the number.
For multiples of both 3 and 5, print "FizzBuzz" instead of the number.